requirements.txt
flask>=2.0.0
gunicorn>=20.1.0
altair
pandas
hugchat>=0.0.8
beautifulsoup4>=4.12.2
openapi-schema-pydantic>=1.2.4
streamlit
python-dotenv>=1.0.0
accelerate
bitsandbytes
gradio
protobuf
scipy
sentencepiece
transformers
requests
glob2
pandas
pydantic==2.5.3
numpy
html-parser
regex
tiktoken
openai[embeddings]
langchain
langchain_community
sentence_transformers
tiktoken
pdfminer.six
openai==0.27.7
faiss-cpu
unstructured
htbuilder
